+++ .pio/libdeps/micro/Metro-Arduino-Wiring/Metro.h
@@ -36,7 +36,7 @@ public:
   Metro();
   Metro(unsigned long interval_millis);
   void interval(unsigned long interval_millis);
-  uint8_t check();
+  uint16_t check();
   void reset();
 	
 private:
