+++ include/CAN_MCP2515.h
@@ -431,7 +431,7 @@ class CAN_MCP2515 : public CANClass
     // Use pin 10 for SPI CS. Allows multiple CAN channels.
     CAN_MCP2515();
     // SPI CS is selectable through sketch. Allows multiple CAN channels.
-    CAN_MCP2515(uint8_t CS_Pin);
+    CAN_MCP2515(uint16_t CS_Pin);
 
     // Initializes CAN communications into Normal mode. Note it also starts SPI communications
     inline void begin (uint32_t bitrate)
@@ -439,28 +439,28 @@ class CAN_MCP2515 : public CANClass
       begin(bitrate, MCP2515_MODE_NORMAL);
     };
     // Initializes CAN communications. Note it also starts SPI communications
-    void begin (uint32_t bitrate, uint8_t mode);
+    void begin (uint32_t bitrate, uint16_t mode);
 
     void end();
     // check if message has been received on any of the buffers
-    uint8_t available();
+    uint16_t available();
     //Receive and display CAN message and allows use of the message structure for easier message handling
     CAN_Frame read();
     // Receive and display any message (J1939, CANopen, CAN)
-    void read(uint32_t *ID, uint8_t *length_out, uint8_t *data_out);
+    void read(uint32_t *ID, uint16_t *length_out, uint16_t *data_out);
 
     void flush();
 
-    uint8_t write(const CAN_Frame&);
+    uint16_t write(const CAN_Frame&);
     // Load and send message. No RTS needed.
-    uint8_t write(uint32_t ID, uint8_t frameType, uint8_t length, uint8_t *data);
+    uint16_t write(uint32_t ID, uint16_t frameType, uint16_t length, uint16_t *data);
 
   private:
-    uint8_t CS; //SPI CS is selectable through sketch
+    uint16_t CS; //SPI CS is selectable through sketch
     void _init();
 
     void reset(); //CAN software reset. Also puts MCP2515 into config mode
-    uint8_t getMode(); // reads CAN mode
+    uint16_t getMode(); // reads CAN mode
     uint32_t getBitrate(); // reads CANspeed
 
     void clearRxBuffers(); // clears all receive buffers
@@ -470,21 +470,21 @@ class CAN_MCP2515 : public CANClass
     // Make setting bitrate and mode only part of the constructor
     void setBitrate(uint32_t bitrate);//sets up CAN bit rate
     void setBitrate16MHz(uint32_t bitrate);
-    void setMode(uint8_t mode);//puts CAN controller in one of five modes
+    void setMode(uint16_t mode);//puts CAN controller in one of five modes
 
-    void writeAddress(uint8_t address, uint8_t value);// writes MCP2515 register addresses
-    uint8_t readAddress(uint8_t address); //reads MCP2515 registers
-    void modifyAddress(uint8_t address, uint8_t mask, uint8_t value); // MCP2515 SPI bit modification commands
+    void writeAddress(uint16_t address, uint16_t value);// writes MCP2515 register addresses
+    uint16_t readAddress(uint16_t address); //reads MCP2515 registers
+    void modifyAddress(uint16_t address, uint16_t mask, uint16_t value); // MCP2515 SPI bit modification commands
 
-    uint8_t readStatus(); //reads several status bits for transmit and receive functions.
-    uint8_t readRXStatus(); //reads receive functions and filter hits
+    uint16_t readStatus(); //reads several status bits for transmit and receive functions.
+    uint16_t readRXStatus(); //reads receive functions and filter hits
 
     // OTHER USEFUL FUNCTIONS
 
     void enableRTSPins (); // Enable hardware request to send (RTS) pins if they are available. It allows messages to be send by driving MCP2515 RTS pins low.
-    void setInterrupts(uint8_t mask, uint8_t writeVal); //Enable/disable interrupts
-    void setMask(uint8_t mask, uint8_t b0, uint8_t b1, uint8_t b2, uint8_t b3); // Set Masks for filters
-    void setFilter(uint8_t filter, uint8_t b0, uint8_t b1, uint8_t b2, uint8_t b3); //Set Receive filters
+    void setInterrupts(uint16_t mask, uint16_t writeVal); //Enable/disable interrupts
+    void setMask(uint16_t mask, uint16_t b0, uint16_t b1, uint16_t b2, uint16_t b3); // Set Masks for filters
+    void setFilter(uint16_t filter, uint16_t b0, uint16_t b1, uint16_t b2, uint16_t b3); //Set Receive filters
 
 };
