+++ src/CAN_SPI_handle.cpp
@@ -11,7 +11,7 @@ void init_SPI()
 
 // -------------------------------------------------
 // Sends through SPI.
-void send_SPI(uint32_t id_1, uint32_t id_2, uint8_t buf_1[], uint8_t buf_2[])
+void send_SPI(uint32_t id_1, uint32_t id_2, uint16_t buf_1[], uint16_t buf_2[])
 {
 
     CAN_Frame msg_1;
@@ -27,14 +27,14 @@ void send_SPI(uint32_t id_1, uint32_t id_2, uint8_t buf_1[], uint8_t buf_2[])
     msg_2.length = 8;
 
     // Load buffers with Temperature array 
-    for (uint8_t INDEX = 0; INDEX < (CHANNELS - 4); INDEX++)
+    for (uint16_t INDEX = 0; INDEX < (CHANNELS - 4); INDEX++)
     {
         msg_1.data[INDEX] = buf_1[INDEX];
         msg_2.data[INDEX] = buf_2[INDEX];
     }
 
 #if DEBUG
-    for (uint8_t INDEX = 0; INDEX < (CHANNELS / 2); INDEX++)
+    for (uint16_t INDEX = 0; INDEX < (CHANNELS / 2); INDEX++)
     {
         Serial.print("Data1 ");
         Serial.print(INDEX);
@@ -49,7 +49,7 @@ void send_SPI(uint32_t id_1, uint32_t id_2, uint8_t buf_1[], uint8_t buf_2[])
     Serial.print("Max 1: ");
     Serial.println(msg_1.data[7]);
 
-    for (uint8_t INDEX = 0; INDEX < (CHANNELS / 2); INDEX++)
+    for (uint16_t INDEX = 0; INDEX < (CHANNELS / 2); INDEX++)
     {
         Serial.print("Data2 ");
         Serial.print(INDEX);
