+++ src/TempSensor.cpp
@@ -7,7 +7,7 @@
 // ---------------------------------------------------------------------
 // Channel Identities in a array
 
-uint8_t portPins[CHANNELS] = {CHANNEL1, CHANNEL2, CHANNEL3, CHANNEL4,
+uint16_t portPins[CHANNELS] = {CHANNEL1, CHANNEL2, CHANNEL3, CHANNEL4,
                               CHANNEL5, CHANNEL6, CHANNEL7, CHANNEL8,
                               CHANNEL9, CHANNEL10, CHANNEL11, CHANNEL12};
 
@@ -50,8 +50,8 @@ void TempSensor::updateTemp()
 
 void TempSensor::updateMinTemp()
 {
-    uint8_t minTempM1 = this->Module_HALF1.senceTemp.temp[CHANNEL1];
-    uint8_t minTempM2 = this->Module_HALF2.senceTemp.temp[CHANNEL1];
+    uint16_t minTempM1 = this->Module_HALF1.senceTemp.temp[CHANNEL1];
+    uint16_t minTempM2 = this->Module_HALF2.senceTemp.temp[CHANNEL1];
     
     for (int CHANNEL = 0; CHANNEL < (CHANNELS / 2); CHANNEL++)
     {
@@ -75,8 +75,8 @@ void TempSensor::updateMinTemp()
 
 void TempSensor::updateMaxTemp()
 {
-    uint8_t maxTempM1 = this->Module_HALF1.senceTemp.temp[CHANNEL1];
-    uint8_t maxTempM2 = this->Module_HALF2.senceTemp.temp[CHANNEL1];
+    uint16_t maxTempM1 = this->Module_HALF1.senceTemp.temp[CHANNEL1];
+    uint16_t maxTempM2 = this->Module_HALF2.senceTemp.temp[CHANNEL1];
     for (int CHANNEL = 0; CHANNEL < (CHANNELS / 2); CHANNEL++)
     {
         if (this->Module_HALF1.senceTemp.temp[CHANNEL] > maxTempM1)
@@ -98,39 +98,41 @@ void TempSensor::updateMaxTemp()
 // -----------------------------------------------------------------
 // Returns the temperatures for a Module half
 
-uint8_t *TempSensor::getTempModuleHALF1()
+uint16_t *TempSensor::getTempModuleHALF1()
 {
     this->Module_HALF1.senceTemp.temp[6] = this->getMinTempModuleHALF1();
     this->Module_HALF1.senceTemp.temp[7] = this->getMaxTempModuleHALF1();
+
     return this->Module_HALF1.senceTemp.temp;
 }
 
-uint8_t *TempSensor::getTempModuleHALF2()
+uint16_t *TempSensor::getTempModuleHALF2()
 {
     this->Module_HALF2.senceTemp.temp[6] = this->getMinTempModuleHALF1();
     this->Module_HALF2.senceTemp.temp[7] = this->getMaxTempModuleHALF2();
+
     return this->Module_HALF2.senceTemp.temp;
 }
 
 // ------------------------------------------------------------------
 // Returns the min and max of each module half
 
-uint8_t TempSensor::getMinTempModuleHALF1()
+uint16_t TempSensor::getMinTempModuleHALF1()
 {
     return this->Module_HALF1.senceTemp.minTemp;
 }
 
-uint8_t TempSensor::getMinTempModuleHALF2()
+uint16_t TempSensor::getMinTempModuleHALF2()
 {
     return this->Module_HALF2.senceTemp.minTemp;
 }
 
-uint8_t TempSensor::getMaxTempModuleHALF1()
+uint16_t TempSensor::getMaxTempModuleHALF1()
 {
     return this->Module_HALF1.senceTemp.maxTemp;
 }
 
-uint8_t TempSensor::getMaxTempModuleHALF2()
+uint16_t TempSensor::getMaxTempModuleHALF2()
 {
     return this->Module_HALF2.senceTemp.maxTemp;
 }
